Move schema management to its own class.
